Add ability for integration flow to build release branches
The flow being if we find we needed to patch an older version we could
create a release branch off the old tag and work on that in the same way
we work on main. Images will be tagged with the branch name and the
branch should be transitioned with the Release Transition workflow to
ensure the image tags in the deployments are updated.

Signed-off-by: Robert Young <robeyoun@redhat.com>